Added outer-join fn
This commit includes the outer-join addition, resulting modification of
the regular join fn and also the needed changes in internal. (build-join)
now accepts a vector arg.

to-fieldlist automatically qualifies its cols.

Changed the compiler to work differently depending on whether it gets a
RTable in the :data field of :joins, or whether its a cross join (:else)
or finally whether its a non join query.