clojure 1.3.0 compatibility
clojure-contrib is deprecated in favor of individual libraries.
c.c.sql is no longer maintained and has some serious bugs that are being
addressed in the new org.clojure/java.jdbc (see for instance
http://dev.clojure.org/jira/browse/JDBC-11)
clojure 1.3.0 requires the addition of :dynamic metadata. All these issues are
taken care in the commit. All tests are passing, regular and integration, using both
clojure 1.2.0 and 1.3.0-beta1. Using clojure 1.3.0 there is a little issue with the
mocking library used for a couple of tests.

The downside is that projects using clojureql and c.c.sql will need to change
namespaces to the new clojure.java.jdbc.

Detail:

* No longer depend on clojure-contrib
** c.c.sql replaced by org.clojure/java.jdbc
** c.c.core replaced by org.clojure/core.incubator
** c.c.mock replaced by org.clojure.contrib/mock